Implemented Changle and CannonAdjust
Set up a system for moving the cannon angle, while also holding it if
the joystick isn't being manipulated. Created methods to enable and
disable the PID controller to allow the catapult angle to hold against
the force of the string. Renamed the DIO RobotMap values for better
redability and understanding.